project 2차

+++ src/member/Member.java
@@ -8,9 +8,9 @@ public class Member {
 	private String pw;
 	private String address;
 	private String phone;
-	private double point;
+	private int point;
 
-	public Member(int memcode, String name, String id, String pw, String address, String phone, double point) {
+	public Member(int memcode, String name, String id, String pw, String address, String phone, int point) {
 		super();
 		this.memcode = memcode;
 		this.name = name;
@@ -89,11 +89,11 @@ public class Member {
 		this.phone = phone;
 	}
 
-	public double getPoint() {
+	public int getPoint() {
 		return point;
 	}
 
-	public void setPoint(double point) {
+	public void setPoint(int point) {
 		this.point = point;
 	}
